# Judge level before call Log()