Test on Windows amd64：go test -v -race -run=None -benchmem -bench=.

New：
```
BenchmarkBufferedFileLogger-6            1000000              2195 ns/op              83 B/op          0 allocs/op
BenchmarkDiscardLogger-6                 1000000              1466 ns/op              83 B/op          0 allocs/op
BenchmarkMultiLevel-6                     100000             15777 ns/op            1321 B/op         25 allocs/op
```

Old：
```
BenchmarkBufferedFileLogger-6            1000000              2191 ns/op              83 B/op          0 allocs/op
BenchmarkDiscardLogger-6                 1000000              1441 ns/op              83 B/op          0 allocs/op
BenchmarkMultiLevel-6                     100000             21552 ns/op            1780 B/op         33 allocs/op
```