@keakon 2018-11-25
把 runtime.Caller() 移出来后确实忘了先判断 level，这里挺希望 Go 有宏之类不影响运行时性能的功能。

PR 中我不确定对 handlers 进行不稳定排序是否会有其他影响，容我先想想。

如果像现在这样没有暴露 SetLevel() 的接口，似乎可以直接把低于 logger 的 handler 丢掉。如果 handlers 为空，可以直接 return。

另一个考量点是不太想把下层的逻辑暴露到上层来，比如所有 formatter 都没有要求输出时间或源码时，生成 record 时就不构造这部分代码。但考虑到一般不会有这样的需求，所以选择了不增加代码的耦合度。