@He11oLx 2018-11-25
考虑 一般都是初始化时根据 ENV 设置 level 一般大于 debug，代码中 debug >> info >> crit。
所以，在 Caller() 之前判断 level 是有必要的。

是否增加 SetLevel，这个不好权衡，各有利弊。

最后一个，提供一个思路。
是否可以，只在 部分 Level 提供 时间或源码，而不是按 formatter ？
```
SetDefaultCallerLevel(WARNLEVEL)

if err!=nil {
    // 此时，可以利用 runtime，获取 行号
    log.Warn("")
}
log.Info("一般的 info 就不获取 行号")
```