@keakon 2018-11-26
考虑到我自己是初始化时设置好日志相关的设定，之后不再修改，所以可以不暴露 SetLevel()。而排序是一次性的，没有太大的性能影响，可以改成 sort.Stable()。

我在调试时经常会加一些 debug log，但又懒得去写 message 的内容，这时候主要根据行号来定位代码走向。